third_party/blink/web_tests/platform/generic/http/tests/messaging/cross-domain-message-event-dispatch-expected.txt

waiting...
Test begins
Events queued
Capture phase: Message1
Bubble phase: Message1
Capture phase: Message2
Bubble phase: Message2
Capture phase: Message3
Bubble phase: Message3
setTimeout1
setTimeout2
